add `datalad_next.shell.shell` context manager
The `shell` context manager
opens a connection to a shell. Typically
the shell is executed remotely via ssh,
but it might also be a local shell, e.g. `bash`.
It allows to send commands to the shell,
read the output of `stdout` and determines
the return code of the command.
When the context is left, the shell is
closed, by closing its `stdin`.

The `shell` context manager is extensible
through the concept of "Response Generators".
New response generator-classes can be used
to support new operating system, differing
command line syntax, or different behavior
of commands on different operating system.
Currently it supports `bash` and `PowerShell`.

The commit also includes a number of
"convenience" functions for Linux and OSX
that can be executed on a `shell`-context.

 - download files
 - upload files
 - delete files

This commit incoroporates
fixes and suggestions from:

Adina Wagner <adina.wagner@t-online.de>
Michael Hanke <michael.hanke@gmail.com>

Many thanks for that.