Enable running `iter_annexworktree()` on just-Git repos
This fixes an obscure

`iterable did not yield matching item for route-in item, cardinality
mismatch?`

error when attempting such a query before.

This fix is a choice. An alternative could be to error out in this case.
I went for this approach, because the test is cheap, and not much more
expensive than catching the previous error in a precise manner.

With this change, the items of the underlying `gitworktree` iterator are
yielded, after funneling them through the `annexworktree` iteration
post-processing to ensure type-alignment and other (future) side-effect.
The outcome is therefore identical to items in Git yielded from an
annex-enabled repository.

Closes: https://github.com/datalad/datalad-next/issues/670