Detect a modification of an added/staged file
Decided to reuse the existing `modification_types` qualifiers.
It is a bit weird that the type enum is called
`GitContainerModificationType`. It is primarily targeting
submodules -- so this makes some sense. But it does not make sense
to install another enum for the same purpose. It should probably
be just `GitModificationType`.