Move away from Git behavior for status reporting
The previous setup used red/green (like git-status) to communicate
the nature of a change. But that was poorly aligned. See
https://github.com/datalad/datalad-next/issues/640

This change does NOT align with Git's behavior. Instead, it acknowledges
even more that DataLad does not work with Git's staging of content in
the index.

Color coding is now purely about the nature of the change (add, delete,
mod, untracked), and not its status with respect to the index.

Closes https://github.com/datalad/datalad-next/issues/640