Report on new submodule commits with `next-status -r mono`
The previous implementation assumed that no submodule items need
to be produced, because individual modification reports from
within the submodules would be sufficient.

However, this missed all cases with new submodule commits, but
no other changes.

This change adjusts the implementation and tests to produce and
verify modification reports for such cases.

Closes https://github.com/datalad/datalad-next/issues/645