support patching of an already loaded class
This commit reloads `cls` in
`datalad.customremotes.main._main` to allow
`cls` to be patched successfully by
`load_extensions()`. This is currently
used for `ORARemote`-patching.

Note: a nicer solution would be to execute:

    from datalad.support.entrypoints import load_extensions
    load_extensions()

as first commands in the `main`-function
of all remote-implementations. But since this
would require changes in datalad-core, it was
not done here.