`set_gitconfig_items_in_env()`: treat `None` as empty string
This violates the promises made in the type annotation of this
function. However, it is necessary (at least for now) in order
to regain compatibility with an (anti-)feature of the DataLad
CLI.

The underlying issue is documented in
https://github.com/datalad/datalad/issues/7589

TL;DR: DataLad promises to be able to unset a git config
item -- but this is not supported by Git, and also DataLad
can only achieve this within a single, top-level process.

With this commit `None` value (e.g. coming from the CLI)
is treated as an empty string. It also ensures that all
values are strings instead of, for example,
`int`.

It adds a paragraph to the doc-string of
`datalad_next.config.utils.set_gitconfig_items_in_env`
to document its behavior if a configuration item
has a `None`-value.

This isssue has been present since the introduction of this
function, but has only been discovered now, because test
setup did not reliably load extensions, hence this datalad-next
code did not always run.

Ping https://github.com/datalad/datalad-next/issues/677