interpret a `None`-env value as empty string
This commit interprets a None environment value
as an empty string. It also ensures that all
values are strings instead of, for example,
`int`.

It adds a paragraph to the doc-string of
`datalad_next.config.utils.set_gitconfig_items_in_env`
to document its behavior if a configuration item
has a `None`-value.

This commit is unrelated to Windows-support,
but became necessary because the datalad-next CI
runs datalad tests now always with the loaded
`datalad_next`-extension, and the previous version
of datalad_next's config code failed the tests.