Relax type assertion from PurePosixPath to PosixPath in RIA patches
The patched code is new patch-code, introduced in
https://github.com/datalad/datalad-next/pull/669

The assertions were added to make sure that no platform paths are
passed around. The Debian maintainer observed non-pure paths coming in.
This may need an investigation, but this changeset maintains the spirit
of the assertions while relaxing them enough to let the tests pass in
a Debian build environment.