chore: remove `git ls-files` filter kludge
This worked around a problem with Git outputing information on files
within the `.git` directory. This was fixed in version 2.25, which
is presently the minimum supported version declared by DataLad.

This also side-steps a problem of a software bug in the kludge handling
that would have led to a crash, if ever triggered.