feat(iter_gitdiff): add pathspec support
Previously, `iter_gitdiff()` did not support path-based evaluation
constraints. This change adds support for that.

Unlike the corresponding implementation in datalad-core, this new
implementation also supports Git's "magic" pathspec. These allow
for more complex expressions, such as matching to all files with
a particular `gitattribute` set or unset.

See https://css-tricks.com/git-pathspecs-and-how-to-use-them/ for a
nice description.

The implementation goes beyond what is supported by Git directly,
and is also capable of pathspec-constraints combined with submodule
recursion.

For the latter, the implementation relies on `GitPathSpec.for_subdir()`.
Some care has been taken to consolidate all pathspec translation
logic in this method, such that fixes and improvement should not
require widespread adjustments in other components.

For now, a basic test is included that covers all new code but one line.

Refs: https://github.com/datalad/datalad-next/issues/598