rf(pathspec): rewrite sub-pattern matching
The new approach may not be much less complex, but it is more correct
and covers more scenarios. It now lives in a standlone function that
performs a purely lexical analysis. It also comes with a dedicated
test collection that is leaner and easier to extend than the previous
one (which remains also).