feat(GitPathSpecs): additional methods to centralize handling
There is a new `any_match_subdir()` that provides the fast-exit test for
a subdir-scope match that is implemented in `iter_submodules()`.
Moreover, there is standard `__eq__()` method to aid identity
comparisions.

The tests of `GitPathSpecs` have been brought to 100% coverage. This
raised a bunch of questions regarding semantics. What is the diffrence
between:

- `GitPathSpecs(None)`
- `GitPathSpecs([])`
- `GitPathSpecs([':'])`

This needs to be documented.