rf(GitPathSpecs): let `for_subdir()` raise when no translation possible
This change is motivated by the need to be able to distinguish
no-constraint scenarios from the inability to translate a constraint
into a particular directory scope -- which typically means that no match
would be possible.

A `None` argument to `GitPathSpecs` is intepreted as no-contraint (i.e.,
a `:` pathspec), which translates to an empty `arglist()` return.

It was confusing that an empty list returned from `for_subdir()`
simultaneously meant 'no-match' rather than 'no-constraint'.

This ambiguity is now resolved with an explicit `ValueError` for the
path that produces a `no-match` result. This enables the implementation
of early-stopping testing in client code -- which is demo'd for
`iter_gitworktree()` in this changeset.