rf(`iterable_subprocess`): removed package and use from `datasalad`
This change introduces a new dependency to the `datasalad` library. This
library provides (updated) implementations of the `iterable_subprocess`
package that was a part of `datalad-next`, and is removed here.

This migration poses the challenge of associated API changes and a new
exception type (`datasalad.runners.CommandError`). This is (for now)
mitigated here by provinding a shim for `iter_subproc()`, the main
entrypoint, that translates the API and exception. This shim is
immediately marked as deprecated.

Future migrations are expected to strengthen the dependency on
`datasalad` further.