doc: exclude memory addresses from documentation
This is achieved by a new patch, because the behavior is caused by
datalad-core's tooling.

The approach taken is to delete memory address statements like
'... at 0x...', and delete them via a regex -- similar to what
sphinx is doing for documentation generation.

The patch is not similar to sphinx, but global, because memory addresses
also have little to no utility for runtime documentation too.

This actually affects only a single usage of a function pointer
used as a default argument.

Closes: https://github.com/datalad/datalad-next/issues/709