fix(delete): set write-permission in `SshUrlOperations.delete`
The method `SshUrlOperations.delete` now attempts to
set write permissions on directories and files that
should be deleted.

The doc-string is amended to describe the behavior of
delete w.r.t. handling non-empty directories and
write-protected directories.

This commit also includes more stderr output in exceptions
that are raised in `SshUrlOperations`.