fix(`gitworktree`): decode bytes before itemization
`decode_bytes()` can yield multiple output chunks for one input chunk
(in the case of decoding errors). This implies that it cannot be
meaningfully used after `itemize()` without threatening the semantics
of the items (i.e., with split-by-line items are no longer unique
lines).

For this reason, this commit changes the order of these helpers in
the `gitworktree()` implementation.

Refs: https://github.com/datalad/datalad-next/issues/740