fix(`decode_bytes`): always backslashreplace when asked to
The previous implementation enabled error handling in decoding only for
the segment of a bytestring that an exception was raised for.

However, it may well be that more decoding errors exist in other parts
of the bytestring. I have a complicated real-world case where this
happens, i.e. raising `UnicodeDecodeError` again, even though
`decode_bytes` was called with `backslash_replace=True`.  Unfortunately
the data is so large that I did not manage to catch the condition
exactly.

It seems to be a needless sophistication to decode some part of the
bytestring with error handling, but not another.

There is a good chance that this patch is badly interacting with the
logic to obtain the next chunk before attempting a decoding again.