test(`test_gitdiff`): disable kwargs type-checking
There is PEP692 which would do this properly, but that is not
in all the Python versions we are using.

The dict-handling that is relevant here has been constrain to
a single function with a few lines in the previous commits.
This should help gaining sufficient confidence that this
particular implementation does not mess with the types.