test(`iter_gitdiff`): new test to show that recursion is broken
Neither is the `yield_tree_items` flag passed properly, nor are
submodules recursed into reliably.

For a superdataset everything is fine. A 1st-level submodule is
reported as of `recursion=no` would be set. There is no reporting
on lower levels of submodules in the general case.

The following commits will address the underlying problems.