fix(`iter_gitdiff`): yield copy of container item when it is also further processed
When asked `iter_gitdiff()` yields a "container item" (directory,
submodule), before it recurses into them. However, during item
postprocessing some properties can be modified further, leading to
unexpected changes that interfere with the recursion.

This change makes sure that a copy of the container item is yielded
in such cases, which prevents this kind of data corruption during
recursion.