feat(`iter_gitdiff`): add support for pathspec-constrained reporting
The implementation largely follows the pattern used in
`iter_gitworktree()`. It concludes a series of refactoring that aligned
both implementations.

A top-level helper is monitored for reports on submodules. If pathspecs
are giving, an additional query, unconstrained by pathspecs, is made to
discover any submodules modified. For each such submodule, the given
pathspecs are ported to the scope of the submodule, and if matching,
a pathspec-constrained is generated for that submodule too.

In this fashion, `iter_gitdiff()` can report on pathspec-constrained
matches, where the constrained does not necessarily match the submodule
root.