rf: introduce `ResultHandler`
`eval_func()` and its dependencies are a spaghetti monster. This commit
introduces a `ResultHandler`, a class that provides functionality to
handle results produced by a command execution.

This idea is to factor out all result processing from result generation
code. Moreover, a caller should eventually be able to provision an
alternative result handler with different implementations for
filtering, and rendering results.