ci: switch to codecov cli on appveyor
This change replaces the legacy codecov uploader on macos with a
specific codecov-cli (the uploader's modern replacement) version that is known
to support several CPUs.
This does not solve the issue that coverage reports are not uploaded to
codecov, but it fixes the issue that the CPU mismatch caused a non-zero
exit code that would lead to a CI failure even when all tests passed.
Fixes https://github.com/datalad/datalad-next/issues/756