fix: stop removing a `logger` attribute from the results
The helper `_execute_command_()` is not the place to decide on the
fate of a results, or to impose any constraints on how it can be
used.

If a logger "does not serialize, or pollutes the output", this is only
relevant, when a result needs to be serialized, or printed.

Any sensible implementation of such functionality will need to have
means to deal with this problem (a logger is not the only source
of complications). Removing it here is therefore premature.