chore: include link to documentation page in project urls
The project.urls are picked up and displayed (with matching icons, if
possible) by pypi.org. DataLad-next has its online documentation under
docs.datalad.org and it is pretty good. Given that we already include
a GitHub link in project.urls anyway (twice, in fact), it could be
better to point Documentation to the online docs, rather than the
GitHub README.