fix: include ZipArchiveOperations in Archivist's handlers
This change equips the archivist special remote with the ability to
oprerate on ZIP files. Now, ZIP is included among the the archive
types checked (if/else) when selecting the right archive handler.

ZipArchiveOperations have been added in v1.1.0 (gh #578), and the
changelog includes the following:

> A new ZipArchiveOperations class added support for ZIP files, and
> enables their use together with the archivist git-annex special
> remote.

However, the ZipArchiveOperations were not used. Archivist only used
TarArchiveOperations, and raised a NotImplementedError if the archive
type was other than tar. This is now addressed.