chore: remove patch that is already included in datalad
The patch to datalad updates '_choose_update_target'
has been proposed as a PR in https://github.com/datalad/datalad/pull/7522/
and was merged Oct. 30th 2023.
As the code is part of datalad already, I see no reason
that this patch needs to be applied or stay in the code base